refactor: move and rename helpers
Split primitives and helpers to their own functions. All primitives now
take only connect request objects. Many helpers were renamed to reduce
confusion with their primitives counterparts.
